Refactored request filter and attribute pipeline

+++ src/SFA.DAS.EAS.Employer_Account.Database/Tables/Membership.sql
@@ -2,7 +2,7 @@
 (
     [AccountId] BIGINT NOT NULL, 
     [UserId] BIGINT NOT NULL, 
-    [RoleId] TINYINT NOT NULL, 
+    [RoleId] INT NOT NULL, 
 	[CreatedDate] DATETIME NOT NULL DEFAULT GETDATE(),
     [ShowWizard] BIT NOT NULL DEFAULT 1, 
     CONSTRAINT [FK_Membership_Account] FOREIGN KEY (AccountId) REFERENCES [employer_account].[Account]([Id]),
